Fix unparsing large xs:hexBinary
In some places, we use an Int for variables referring to the length in
bits of xs:hexBinary data. This means we can only unparse MaxInt bits of
hexBinary data, which is only 256MB. This modifies hex binary unparsing
to use Long's in these cases to ensure we can support large hex binary
blobs. The maxLenghHexBinaryInBits tunable is also checked during
unparsing, which ensures we won't try to unparse more than 2GB of
hexBinary data, which is a limitation of Java byte arrays.

DAFFODIL-2532

+++ daffodil-io/src/main/scala/org/apache/daffodil/io/DataOutputStream.scala
@@ -196,7 +196,7 @@ trait DataOutputStream extends DataStreamCommon
    * It is a usage error if bitLengthFrom1 is not greater than or equal to 1.
    *
    */
-  def putByteArray(ba: Array[Byte], bitLengthFrom1: Int, finfo: FormatInfo, ignoreByteOrder: Boolean = false): Boolean
+  def putByteArray(ba: Array[Byte], bitLengthFrom1: Long, finfo: FormatInfo, ignoreByteOrder: Boolean = false): Boolean
 
   /**
    * Float and Double
